Moved accountManagement to server
