+++ frog/imports/client/App/index.js
@@ -26,7 +26,6 @@ import { LocalSettings } from '/imports/api/settings';
 import WikiRouter from '../Wiki/WikiRouter';
 import SingleActivity from '../SingleActivity';
 import { connection } from './connection';
-import { loginUser } from '../../api/accountManagement';
 
 const TeacherContainer = Loadable({
   loader: () => import('./TeacherContainer'),
