+++ frog/server/main.js
@@ -28,7 +28,7 @@ import dashboardSubscription from './dashboardSubscription';
 import './getLogMethods';
 import { activityTypesObj } from '../imports/activityTypes';
 import initExternalOperators from './externalOperators';
-import { createAccount } from '../imports/api/accountManagement';
+import { createAccount } from './accountManagement';
 import './redis';
 
 require('util').inspect.defaultOptions.depth = null;
